code was cleaned before norming, everything seems to be working fine

+++ client.c
@@ -1,6 +1,6 @@
 #include "minitalk.h"
 
-#include <stdio.h>
+// #include <stdio.h>
 
 
 void	my_exit(int status, char *exit_str)
@@ -9,7 +9,7 @@ void	my_exit(int status, char *exit_str)
 	exit(status);
 }
 
-int	string_isdigit(char *str)
+int	str_isdigit(char *str)
 {
 	while (*str)
 	{
@@ -26,13 +26,12 @@ void	send_null(pid_t server_pid)
 
 	if (bits)
 	{
-		kill(server_pid, SIGUSR2);
+		if (kill(server_pid, SIGUSR2) < 0)
+			my_exit(EXIT_FAILURE, ERR_SENDSIG);
 		bits--;
 	}
 	if (!bits)
-	{
 		my_exit(EXIT_SUCCESS, "Message was successfully sent");
-	}
 }
 
 void	send_message(char *str, pid_t pid)
@@ -41,55 +40,25 @@ void	send_message(char *str, pid_t pid)
 	static int		bits = 8;
 	static int		server_pid;
 
-	if (str) {
-		message = str; }
+	if (str)
+		message = str;
 	if (pid)
 		server_pid = pid;
-
 	if (!bits)
 	{
 		message++;
-		// printf("%s\n", message);
 		bits = 8;
 	}
-
-	if (*message)
-	{
-		// if (bits)
-		// {
-			if (*message & (1 << (bits - 1)))
-			{
-				if (kill(server_pid, SIGUSR1) < 0)
-					my_exit(EXIT_FAILURE, ERR_SIG);
-				// ft_putchar('1');
-				// printf("%d bits\n", bits);
-				// usleep(50);
-			}
-			else
-			{
-				if (kill(server_pid, SIGUSR2) < 0)
-					my_exit(EXIT_FAILURE, ERR_SIG);
-				// ft_putchar('0');
-				// printf("%d bits\n", bits);
-				// usleep(50);
-			}
-			bits--;
-		// }
-		// if (!bits)
-		// {
-		// 	message++;
-		// 	// printf("%s\n", message);
-		// 	bits = 8;
-		// }
-	}
-	else
-	{
+	if (!(*message))
 		send_null(server_pid);
-		// kill(server_pid, SIGUSR2);
-		// usleep(50);
-		// bits--;
+	if (*message && (*message & (1 << (bits - 1))))
+	{
+		if (kill(server_pid, SIGUSR1) < 0)
+			my_exit(EXIT_FAILURE, ERR_SENDSIG);
 	}
-	// printf("%d bits\n", bits);
+	else if (*message && kill(server_pid, SIGUSR2) < 0)
+		my_exit(EXIT_FAILURE, ERR_SENDSIG);
+	bits--;
 }
 
 void	signal_handler(int sig, siginfo_t *siginfo, void *context)
@@ -98,32 +67,21 @@ void	signal_handler(int sig, siginfo_t *siginfo, void *context)
 	(void)siginfo;
 	if (sig == SIGUSR1)
 		send_message(NULL, 0);
-	// else if (sig == SIGUSR2)
-	// 	exit(0);
+	else if (sig == SIGUSR2)
+		my_exit(EXIT_FAILURE, ERR_FROM_SERV);
 }
 
 int	main(int argc, char **argv)
 {
-	char			*message;
-	pid_t			server_pid;
 	struct sigaction	sa;
 
-	if (!(argc == 3 && string_isdigit(argv[1])))
+	if (!(argc == 3 && str_isdigit(argv[1])))
 		my_exit(EXIT_FAILURE, ERR_WRARG);
-
-	server_pid = ft_atoi(argv[1]);
-	message = argv[2];
-
 	sa.sa_sigaction = signal_handler;
 	sa.sa_flags = SA_SIGINFO;
-
-	sigaction(SIGUSR1, &sa, NULL);
-	sigaction(SIGUSR2, &sa, NULL);
-	send_message(argv[2], server_pid);
-
-	// ft_putstr_fd("Hi from client, my PID is ", 1);
-	// ft_putnbr_fd(getpid(), 1);
-	// ft_putchar_fd('\n', 1);
+	if (sigaction(SIGUSR1, &sa, NULL) || sigaction(SIGUSR2, &sa, NULL))
+		my_exit(EXIT_FAILURE, ERR_SIGACT);
+	send_message(argv[2], ft_atoi(argv[1]));
 	while (true)
 		pause();
 	return (0);
